requirements.txt
altair==5.2.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
entrypoints==0.4
Faker==22.2.0
favicon==0.7.0
filelock==3.18.0
fonttools==4.47.2
gitdb==4.0.11
GitPython==3.1.41
google==3.0.0
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-aiplatform==1.39.0
google-cloud-bigquery==3.16.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-genai==1.18.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.16.0
htbuilder==0.6.2
httpcore==1.0.9
httpx==0.28.1
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.2.0
narwhals==1.41.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
platformdirs==4.3.8
plotly==6.1.2
prometheus_client==0.22.0
proto-plus==1.23.0
protobuf==4.25.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.10.1
pymdown-extensions==10.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
st-annotated-text==4.0.1
st-pages==0.4.5
streamlit==1.30.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-chat==0.0.2.2
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-option-menu==0.3.12
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.0
tornado==6.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
websockets==15.0.1
zipp==3.17.0
